feat: use case builder toggle functionality & refactor huge components to divided small components
