+++ packages/cdk/lib/construct/web.ts
@@ -38,6 +38,7 @@ export interface WebProps {
   hostName?: string;
   domainName?: string;
   hostedZoneId?: string;
+  useCaseBuilderEnabled: boolean;
 }
 
 export class Web extends Construct {
@@ -186,6 +187,8 @@ export class Web extends Construct {
         VITE_APP_SAML_COGNITO_FEDERATED_IDENTITY_PROVIDER_NAME:
           props.samlCognitoFederatedIdentityProviderName.toString(),
         VITE_APP_AGENT_NAMES: JSON.stringify(props.agentNames),
+        VITE_APP_USE_CASE_BUILDER_ENABLED:
+          props.useCaseBuilderEnabled.toString(),
       },
     });
