+++ packages/cdk/lib/generative-ai-use-cases-stack.ts
@@ -14,6 +14,7 @@ import { CfnWebACLAssociation } from 'aws-cdk-lib/aws-wafv2';
 import * as cognito from 'aws-cdk-lib/aws-cognito';
 import { ICertificate } from 'aws-cdk-lib/aws-certificatemanager';
 import { Agent, PromptFlow } from 'generative-ai-use-cases-jp';
+import { UseCaseBuilder } from './construct/use-case-builder';
 
 const errorMessageForBooleanContext = (key: string) => {
   return `${key} の設定でエラーになりました。原因として考えられるものは以下です。
@@ -72,6 +73,9 @@ export class GenerativeAiUseCasesStack extends Stack {
 
     const guardrailEnabled: boolean =
       this.node.tryGetContext('guardrailEnabled') || false;
+    const useCaseBuilderEnabled: boolean = this.node.tryGetContext(
+      'useCaseBuilderEnabled'
+    )!;
 
     if (typeof ragEnabled !== 'boolean') {
       throw new Error(errorMessageForBooleanContext('ragEnabled'));
@@ -95,6 +99,10 @@ export class GenerativeAiUseCasesStack extends Stack {
       );
     }
 
+    if (typeof useCaseBuilderEnabled !== 'boolean') {
+      throw new Error(errorMessageForBooleanContext('useCaseBuilderEnabled'));
+    }
+
     const auth = new Auth(this, 'Auth', {
       selfSignUpEnabled,
       allowedIpV4AddressRanges: props.allowedIpV4AddressRanges,
@@ -108,7 +116,6 @@ export class GenerativeAiUseCasesStack extends Stack {
       userPool: auth.userPool,
       idPool: auth.idPool,
       table: database.table,
-      useCaseBuilderTable: database.useCaseBuilderTable,
       agents: props.agents,
       guardrailIdentify: props.guardrailIdentifier,
       guardrailVersion: props.guardrailVersion,
@@ -161,6 +168,7 @@ export class GenerativeAiUseCasesStack extends Stack {
       hostName: props.hostName,
       domainName: props.domainName,
       hostedZoneId: props.hostedZoneId,
+      useCaseBuilderEnabled,
     });
 
     if (ragEnabled) {
@@ -187,6 +195,14 @@ export class GenerativeAiUseCasesStack extends Stack {
 
       // File API から data source の Bucket のファイルをダウンロードできるようにする
       api.allowDownloadFile(props.knowledgeBaseDataSourceBucketName!);
+    }
+
+    if (useCaseBuilderEnabled) {
+      new UseCaseBuilder(this, 'UseCaseBuilder', {
+        userPool: auth.userPool,
+        api: api.api,
+        useCaseBuilderTable: database.useCaseBuilderTable,
+      });
     }
 
     new Transcribe(this, 'Transcribe', {
@@ -283,6 +299,10 @@ export class GenerativeAiUseCasesStack extends Stack {
 
     new CfnOutput(this, 'AgentNames', {
       value: JSON.stringify(api.agentNames),
+    });
+
+    new CfnOutput(this, 'UseCaseBuilderEnabled', {
+      value: useCaseBuilderEnabled.toString(),
     });
 
     this.userPool = auth.userPool;
